xostools: JACK
jack doesn't help, so let's scream on jack.

Do it like:

JACK

In that way you screamed on him really loud and he will stop derping around and work again.

For noobs / naives / people without sense of humor:

The new command JACK allows you to workaround the problem with the jack compiler.
The problem is that you can build using it once without any problem, but the second time you try it it will get stuck in building Android SDK stubs (android_stubs....) forever and you would need to restart your machine.

This simple command allows you to quickly give a kick on jack so that he continues working for you.

Maybe the reason why this doesn't work is because Apple removed the jack, hence it doesn't want to work properly anymore.

Change-Id: I2c1fdedfa60188648ce8b8c1f60b8ffd4c51c990
Signed-off-by: Simao Gomes Viana <xdevs23@outlook.com>