Fix ninja and ccache not being found
which throws an error if a binary is not found which breaks the build. Use command -v as it is safer.

Also, switch from ?= to := as it was causing ninja not to be found even with command -v. We'll print what binary the user is using so that there is no confusion either.

This also reverts 6afa51812bcf404597ba149b3c0285cf06d11ebb since whatever ccache you got on local
will be better than Google's outdated AF binary for sure

Change-Id: Ic78eb71d06357523bbe483e30898132747d18080
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: Simao Gomes Viana <xdevs23@outlook.com>